+++ config/autoload/templates.global.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * @license    https://opensource.org/licenses/BSD-3-Clause New BSD License
+ * @copyright  (c) 2017-2018, jailgreen jukka@jahlgren.eu
+ */
+
+declare(strict_types=1);
+
+return [
+    'templates' => [
+        'paths' => [
+            'error'  => ['templates/error'],
+            'layout' => ['templates/layout'],
+        ],
+    ],
+];