+++ CHANGELOG.md
@@ -2,16 +2,17 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 1.0.0 - TBD
+## 1.0.0-dev - 2018-05-22
 
 ### Added
 
+- tests for ConfigProvider.
 - docheader and phpstan.
 - expressive skeleton.
 
 ### Changed
 
-- Nothing.
+- Moved `templates` to project root.
 
 ### Deprecated